requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
device-detector==5.0.1
Django==4.2.4
django-allauth==0.54.0
django-user-agents==0.4.0
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
frozenlist==1.4.0
geoip2==4.7.0
httpagentparser==1.9.5
idna==3.4
maxminddb==2.4.0
multidict==6.0.4
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
ua-parser==0.18.0
urllib3==2.0.4
user-agents==2.2.0
yarl==1.9.2
